backend/requirements.txt
backcall==0.2.0
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.1
colorama==0.4.4
dbml-sqlite==0.3.3
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
et-xmlfile==1.1.0
Flask==2.0.2
google-api-core==2.2.1
google-api-python-client==2.28.0
google-auth==2.3.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
greenlet==1.1.1
gspread==4.0.1
gunicorn==20.1.0
httplib2==0.20.1
idna==3.3
ipykernel==6.4.2
ipython==7.29.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
jupyter-client==7.0.6
jupyter-core==4.9.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
numpy==1.21.1
oauth2client==4.1.3
oauthlib==3.1.1
openpyxl==3.0.7
pandas==1.3.1
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.21
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydbml==0.3.4
Pygments==2.10.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.1
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.23
tornado==6.1
traitlets==5.1.1
uritemplate==4.1.1
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
dependency-injector
bootstrap-flask
